refactor: module structures
