+++ src/openaivec/__init__.py
@@ -1,14 +1,13 @@
-from .embeddings import AsyncBatchEmbeddings, BatchEmbeddings
-from .model import PreparedTask
-from .prompt import FewShotPrompt, FewShotPromptBuilder
-from .responses import AsyncBatchResponses, BatchResponses
+from ._embeddings import AsyncBatchEmbeddings, BatchEmbeddings
+from ._model import PreparedTask
+from ._prompt import FewShotPromptBuilder
+from ._responses import AsyncBatchResponses, BatchResponses
 
 __all__ = [
     "AsyncBatchEmbeddings",
     "AsyncBatchResponses",
     "BatchEmbeddings",
     "BatchResponses",
-    "FewShotPrompt",
     "FewShotPromptBuilder",
     "PreparedTask",
 ]
