+++ src/openaivec/_log.py
