+++ src/openaivec/_model.py
