+++ src/openaivec/_optimize.py
