+++ src/openaivec/_serialize.py
