+++ src/openaivec/_util.py
