+++ src/openaivec/task/customer_support/intent_analysis.py
@@ -8,7 +8,7 @@ Example:
 
     ```python
     from openai import OpenAI
-    from openaivec.responses import BatchResponses
+    from openaivec._responses import BatchResponses
     from openaivec.task import customer_support
 
     client = OpenAI()
@@ -61,7 +61,7 @@ from typing import List, Literal
 
 from pydantic import BaseModel, Field
 
-from openaivec.model import PreparedTask
+from openaivec._model import PreparedTask
 
 __all__ = ["intent_analysis"]
