+++ src/openaivec/task/nlp/morphological_analysis.py
@@ -9,7 +9,7 @@ Example:
 
     ```python
     from openai import OpenAI
-    from openaivec.responses import BatchResponses
+    from openaivec._responses import BatchResponses
     from openaivec.task import nlp
 
     client = OpenAI()
@@ -53,7 +53,7 @@ from typing import List
 
 from pydantic import BaseModel, Field
 
-from openaivec.model import PreparedTask
+from openaivec._model import PreparedTask
 
 __all__ = ["MORPHOLOGICAL_ANALYSIS"]
