+++ src/openaivec/task/nlp/sentiment_analysis.py
@@ -8,7 +8,7 @@ Example:
 
     ```python
     from openai import OpenAI
-    from openaivec.responses import BatchResponses
+    from openaivec._responses import BatchResponses
     from openaivec.task import nlp
 
     client = OpenAI()
@@ -52,7 +52,7 @@ from typing import List, Literal
 
 from pydantic import BaseModel, Field
 
-from openaivec.model import PreparedTask
+from openaivec._model import PreparedTask
 
 __all__ = ["SENTIMENT_ANALYSIS"]
