+++ tests/test_provider.py
@@ -4,7 +4,7 @@ import warnings
 
 from openai import AsyncAzureOpenAI, AsyncOpenAI, AzureOpenAI, OpenAI
 
-from openaivec.provider import provide_async_openai_client, provide_openai_client, reset_environment_registrations
+from openaivec._provider import provide_async_openai_client, provide_openai_client, reset_environment_registrations
 
 
 class TestProvideOpenAIClient(unittest.TestCase):
@@ -365,7 +365,7 @@ class TestAzureV1ApiWarning(unittest.TestCase):
 
     def test_check_azure_v1_api_url_no_warning_for_v1_url(self):
         """Test that v1 API URLs don't trigger warnings."""
-        from openaivec.provider import _check_azure_v1_api_url
+        from openaivec._provider import _check_azure_v1_api_url
 
         v1_urls = [
             "https://myresource.services.ai.azure.com/openai/v1/",
@@ -381,7 +381,7 @@ class TestAzureV1ApiWarning(unittest.TestCase):
 
     def test_check_azure_v1_api_url_warning_for_legacy_url(self):
         """Test that legacy API URLs trigger warnings."""
-        from openaivec.provider import _check_azure_v1_api_url
+        from openaivec._provider import _check_azure_v1_api_url
 
         legacy_urls = [
             "https://myresource.services.ai.azure.com/",
